+++ core/src/main/java/gir/io/FileUtil.java
@@ -4,6 +4,7 @@ import gir.GirException;
 import gir.Task;
 import java.io.File;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -151,7 +152,7 @@ public final class FileUtil
   public static void write( @Nonnull final String path, @Nonnull final String content )
     throws IOException
   {
-    write( path, content.getBytes() );
+    write( path, content.getBytes( StandardCharsets.US_ASCII ) );
   }
 
   /**