+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.09](https://github.com/realityforge/gir/tree/v0.09) (2019-05-02)
 [Full Changelog](https://github.com/realityforge/gir/compare/v0.08...v0.09)